Fix first loss cover profit distribution
We are incorrectly terminating the profit distribution loop for
first loss covers early because the profit check is in the loop
condition itself, not the loop body, which means if the profit
for one particular first loss cover is 0, then none of the subsequent
first loss cover would get the chance to have the profit distributed
to them.

The fix is to put the value check in the loop body, and skip that
particular first loss cover if the profit is 0.