Move access control functions exclusively used by TrancheVault out
This PR moves access control functions exclusively used by TrancheVault
out of PoolConfig and into TrancheVault, in order to reduce the size of
PoolConfig, which was > 21KB before the change. We are also creating
a new generic error called `notAuthorizedCaller` for function calls violating access control rules, and replaced some of the access control errors with the new generic error, since the benefit of having one error per
violation seems not so useful.