Remove `Payment` and `PaymentStatus` structs
According to my discussion with Bin, the `Payment` struct was initially intended for real-time PnL tracking, which we are no longer doing. I checked that the struct is only used in `_makePayment`, but it's only written to, not read from. So let's remove the struct and simplify the code flow.