Add additional calendar functions.
The two functions are:
1. `getNextDueDate()` that works with the new pay period durations.
2. `getNumPeriodsPassed()` that counts the number of periods passed between two dates for the given pay period duration.

The current `getNextDueDate()` function returns both the next due date and the number of periods passed since the last due date. However, given that a bill may start mid-period, we need to count the number of days passed for the purpose of calculating the yield and principal due. `getDaysDiff()`, which was added in previous PRs, satisfies this requirements.

There are currently two callers of `getNextDueDate()`, and one of them only cares about the due date, not the number of periods passed. So it seems appropriate to have separate functions for returning the next due date, the number of days passed, and the number of periods passed. This makes sure that:
(1) The caller only gets what it wants, which saves gas since no additional calculation is done.
(2) The individual functions can be made more generic and easier to test.

If we later discover that we need a function that returns all three at once, we can simply add that function.