Credit contract updates for the new `PayPeriodDuration`
1. Update `CreditConfig` to use the new `PayPeriodDuration` enum
2. Record the maturity date in a separate mapping
3. Update principal, yield and late fee calculation using the new calendar functions

Note that:

(1) In order to limit the size of this change, we are assigning `CreditConfig.periodDuration` to monthly intentionally, so that we don't have to simultaneously update the LP side and bloat the size of this change. The LP side will be updated in a separate PR.
(2) Most of the `CreditLine` tests are commented out for now since we need to update them later anyways.