Remove calendar related credit calculation to CreditDueManager
There is one last place in the `Credit` contract where we used the specific day-based calculation for yield and principal due that's not in `CreditDueManager`. This PR moves that last piece into `CreditDueManager` so that we can easily substitute day-based calculation with second-based calculation for other use cases in the future.