Fix variable shadowing
The input param `coveredLoss` for `_calcLossRecovery()` shadows the storage variable `coveredLoss`. Since the internal function has trivial amount of logic, is called only once, and the other var names are quite confusing, let's get rid of the internal function and put the logic in the caller instead. Also used `Math.min` to make the intention of the code clearer.