Remove `EAServiceAccount`
In V1, `EAServiceAccount` was used by the `EAVerse` service to write underwiritng decisions from the EA of each pool to the contract. But in V2, since a lot of the EA decisions are carried out manually, a centralized `EAVerse` service no longer makes sense, and individual EAs should be able to interact with the contract directly to apply their decisions. Hence, instead of checking whether the `msg.sender` is the `EAServiceAccount`, check whether it's the EA's address instead.