Only require min deposit amount in `depositCover`
Do not enforce min deposit amount in `depositCoverFor()` since the function is used by the pool fee manager to deposit on behalf of pool admins when investing in first loss covers, and the amount is not guaranteed to be always greater than the min deposit amount. So let's enforce the `minDepositAmount` on proactive deposits from providers only.