Wrap `safeTransfer` in an external function
USDT has a non-standard implementation for ERC20 where they do not return anything on transfer success, so the current way of error handling would not work correctly. So we need to wrap `safeTransfer()` in an external function to make it work. It's admittedly a hacky solution but may be the only way to make it work.